+++ src/nanoarrow/ipc/files_test.cc
