+++ src/nanoarrow/ipc/reader.c
