+++ src/nanoarrow/ipc/reader_test.cc
